// src/app/api/model/createReclamationDto.ts
/**
 * GestionRFQ API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ReclamationType } from './reclamationType';


export interface CreateReclamationDto { 
    title?: string | null;
    message?: string | null;
    type?: ReclamationType;
}
export namespace CreateReclamationDto {
}